test: unary unit tests
